server/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-debug-toolbar==6.2.0
celery==5.6.2
django-celery-results==2.6.0
django-redis==6.0.0
redis==7.3.0
Pillow==12.1.0
